requirements.txt
aiogram>=3.4.0
aiosqlite>=0.19.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tzlocal>=5.0
pytz>=2023.3

